requirements.txt
streamlit
pandas
scikit-surprise